requirements.txt
aniso8601==7.0.0
blinker==1.4
boto3==1.9.215
botocore==1.12.215
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-RESTful==0.3.7
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mysql-connector==2.2.9
passlib==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
